key-derivation-server/requirements.txt
cog
python-dotenv
bip-utils>=2.9.0 